Remove dead reverse registry and protocol surfaces
