+++ packages/protocol/src/index.ts
@@ -8,6 +8,7 @@ export * from "./metadata.js";
 export * from "./network.js";
 export * from "./requests.js";
 export * from "./storage.js";
+export * from "./state.js";
 export * from "./snapshots.js";
 export * from "./capabilities.js";
 export * from "./cloud.js";
@@ -25,7 +26,6 @@ export * from "./operations.js";
 export * from "./rest.js";
 export * from "./mcp.js";
 export * from "./scripts.js";
-export * from "./reverse.js";
 export * from "./interaction.js";
 export * from "./semantic.js";
 export * from "./computer-use-bridge.js";
