+++ tests/opensteer/filesystem-root.test.ts
@@ -102,18 +102,6 @@ describe("Phase 3 filesystem root", () => {
       "registry/request-plans/indexes/",
       "registry/request-plans/indexes/by-key/",
       "registry/request-plans/records/",
-      "registry/reverse-cases/",
-      "registry/reverse-cases/indexes/",
-      "registry/reverse-cases/indexes/by-key/",
-      "registry/reverse-cases/records/",
-      "registry/reverse-packages/",
-      "registry/reverse-packages/indexes/",
-      "registry/reverse-packages/indexes/by-key/",
-      "registry/reverse-packages/records/",
-      "registry/reverse-reports/",
-      "registry/reverse-reports/indexes/",
-      "registry/reverse-reports/indexes/by-key/",
-      "registry/reverse-reports/records/",
       "traces/",
       "traces/runs/",
       "workspace.json",
